Fix RedModInstaller tests due to the changes in installation structure.
